replace popup by toggle in prediction history list
