+++ frontend/src/components/PredictionHistoryDrawer.vue
@@ -32,42 +32,74 @@
     </div>
 
     <v-list lines="two">
-      <v-list-item
-        v-for="item in history"
-        :key="item.query_id"
-        class="history-item"
-        :disabled="viewingId === item.query_id"
-        @click="viewPrediction(item)"
-      >
-        <template #title>
-          <span class="item-title">{{ formatLabel(item) }}</span>
-        </template>
+      <template v-for="item in history" :key="item.query_id">
+        <v-list-item
+          class="history-item"
+          :disabled="viewingId === item.query_id"
+          @click="viewPrediction(item)"
+        >
+          <template #title>
+            <span class="item-title">{{ formatLabel(item) }}</span>
+          </template>
 
-        <template #subtitle>
-          <span>{{ formatDate(item.created_at) }}</span>
-        </template>
+          <template #subtitle>
+            <span>{{ formatDate(item.created_at) }}</span>
+          </template>
 
-        <template #append>
-          <v-btn
-            icon="mdi-information-outline"
-            variant="text"
-            size="small"
-            @click.stop="showDetails(item)"
-          />
-          <v-btn
-            icon="mdi-tray-arrow-down"
-            variant="text"
-            size="small"
-            @click.stop="exportPrediction(item)"
-          />
-        </template>
-      </v-list-item>
+          <template #append>
+            <v-btn
+              icon="mdi-information-outline"
+              variant="text"
+              size="small"
+              @click.stop="toggleDetails(item)"
+            />
+            <v-btn
+              icon="mdi-tray-arrow-down"
+              variant="text"
+              size="small"
+              @click.stop="exportPrediction(item)"
+            />
+          </template>
+        </v-list-item>
+
+        <v-expand-transition>
+          <div v-if="expandedId === item.query_id" class="details-panel">
+            <div class="detail-row">
+              <span class="detail-label">Prediction ID</span>
+              <span class="detail-value">{{ item.query_id }}</span>
+            </div>
+            <div class="detail-row">
+              <span class="detail-label">Model</span>
+              <span class="detail-value">{{ item.model_name || 'Unknown' }}</span>
+            </div>
+            <div v-if="isZeroShot(item)" class="detail-row">
+              <span class="detail-label">Keyword</span>
+              <span class="detail-value">{{ resolveKeyword(item) || 'Unknown' }}</span>
+            </div>
+            <div class="detail-row">
+              <span class="detail-label">Coordinates</span>
+              <div class="bbox-coords">
+                <span>N {{ item.bbox.max_lat.toFixed(5) }}</span>
+                <span>S {{ item.bbox.min_lat.toFixed(5) }}</span>
+                <span>E {{ item.bbox.max_lon.toFixed(5) }}</span>
+                <span>W {{ item.bbox.min_lon.toFixed(5) }}</span>
+              </div>
+            </div>
+            <div class="detail-row">
+              <span class="detail-label">Area</span>
+              <span class="detail-value">{{ formatArea(computeBboxArea(item.bbox)) }}</span>
+            </div>
+          </div>
+        </v-expand-transition>
+      </template>
     </v-list>
   </v-navigation-drawer>
 </template>
 
 <script setup>
 import { ref } from 'vue'
+import { Polygon } from 'ol/geom'
+import { getArea } from 'ol/sphere'
 import { useMapStore } from '@/stores/map'
 
 const API_BASE_URL = '/api/segmentation' // backend API
@@ -78,6 +110,7 @@ const history = ref([])
 const loading = ref(false)
 const error = ref(null)
 const viewingId = ref(null)
+const expandedId = ref(null)
 
 
 function toggleDrawer() {
@@ -147,9 +180,50 @@ async function fetchGeoJSONResponse(result) {
   return response
 }
 
-// open the details popup for a specific past prediction
-function showDetails(item) {
-  mapStore.openDetailsDialog(item)
+// toggle the inline details panel for a specific past prediction
+function toggleDetails(item) {
+  expandedId.value = expandedId.value === item.query_id ? null : item.query_id
+}
+
+function computeBboxArea(bbox) {
+  if (!bbox) return null
+
+  const ring = [
+    [bbox.min_lon, bbox.min_lat],
+    [bbox.max_lon, bbox.min_lat],
+    [bbox.max_lon, bbox.max_lat],
+    [bbox.min_lon, bbox.max_lat],
+    [bbox.min_lon, bbox.min_lat],
+  ]
+
+  const polygon = new Polygon([ring])
+  return getArea(polygon, { projection: 'EPSG:4326' })
+}
+
+function formatArea(sqm) {
+  if (sqm == null) return ''
+  return sqm > 1_000_000
+    ? `${(sqm / 1_000_000).toFixed(2)} km²`
+    : `${Math.round(sqm)} m²`
+}
+
+function isZeroShot(item) {
+  return Boolean(item.prediction_type?.toLowerCase().includes('zero_shot'))
+}
+
+// Older predictions didn't persist `keywords`; fall back to the summary
+// text, which already embeds the keyword (e.g. "Found 113 house polygons/clusters").
+function resolveKeyword(item) {
+  if (item.keywords && item.keywords.length) return item.keywords.join(', ')
+  if (!item.summary) return null
+
+  const forMatch = item.summary.match(/\bfor\s+(.+)$/i)
+  if (forMatch) return forMatch[1]
+
+  const countMatch = item.summary.match(/^Found\s+\d+\s+(.+?)\s+polygons\/clusters/i)
+  if (countMatch) return countMatch[1]
+
+  return null
 }
 
 // open the export dialog for a specific past prediction
@@ -238,5 +312,37 @@ function formatDate(isoString) {
 
 .history-item {
   cursor: pointer;
+}
+
+.details-panel {
+  padding: 8px 16px 12px;
+  background: rgba(0, 0, 0, 0.03);
+}
+
+.detail-row {
+  display: flex;
+  justify-content: space-between;
+  align-items: baseline;
+  gap: 12px;
+  padding: 4px 0;
+  font-size: 13px;
+}
+
+.detail-label {
+  font-weight: 600;
+  color: rgba(0, 0, 0, 0.6);
+  white-space: nowrap;
+}
+
+.detail-value {
+  text-align: right;
+  word-break: break-all;
+}
+
+.bbox-coords {
+  display: flex;
+  flex-direction: column;
+  align-items: flex-end;
+  gap: 2px;
 }
 </style>
