+++ frontend/src/views/Home.vue
@@ -9,7 +9,6 @@
     </div>
     <PredictionHistoryDrawer />
     <ExportDialog />
-    <PredictionDetailsDialog />
   </div>
 </template>
 
@@ -21,7 +20,6 @@ import LoadingOverlay from '../components/LoadingOverlay.vue'
 import ErrorOverlay from '../components/ErrorOverlay.vue'
 import Legend from '../components/Legend.vue'
 import ExportDialog from '../components/ExportDialog.vue'
-import PredictionDetailsDialog from '../components/PredictionDetailsDialog.vue'
 </script>
 
 <style scoped>
